server/requirements.txt

# Core Django and REST Framework
Django>=4.2.7
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
# Database
psycopg2-binary>=2.9.9
# Celery and Background Tasks
celery>=5.3.4
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
redis>=5.0.1
# API Documentation
drf-yasg>=1.21.7
# Additional Utilities
python-dotenv>=1.0.0

# Development and Debugging
django-extensions>=3.2.3
werkzeug>=3.0.1

# Production requirements
gunicorn>=21.2.0
whitenoise>=6.5.0

# Monitoring and Logging
sentry-sdk>=1.31.0

# Optional: for advanced features
pillow>=10.0.1  # Image processing


requests>=2.31.0
beautifulsoup4>=4.12.2
gunicorn>=21.2.0

# AI and ML Dependencies
numpy>=1.20.0
scikit-learn>=1.0.0
requests>=2.25.0
python-dotenv>=1.0.0

# Data Processing
pandas>=1.3.0

# Caching
django-redis>=5.0.0

# Security
cryptography>=3.4.0
pyOpenSSL>=20.0.0

# PDF Generation
weasyprint>=53.0
reportlab>=3.6.0

# File handling
python-magic>=0.4.0  # For file type detection

# For HTML handling in exports

lxml>=4.6.0

openai>=1.0.0

# O'Auth

django-oauth-toolkit==1.7.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
requests-oauthlib==1.3.1